+++ infra/docker/traffic_server/docker.nix
@@ -1,12 +1,15 @@
 let
 	pkgs = import ../../nix/common/pkgs.nix;
+	custom_trafficserver = pkgs.trafficserver.override {
+		withLuaJIT = true;
+	};
 	image = pkgs.dockerTools.buildImage {
 		name = "apache-traffic-server";
 		tag = "latest";
 
 		copyToRoot = pkgs.buildEnv {
 			name = "image-root";
-			paths = [ pkgs.trafficserver ];
+			paths = [ custom_trafficserver ];
 			pathsToLink = [ "/bin" ];
 		};
 
@@ -18,7 +21,7 @@ let
 		'';
 
 		config = {
-			Entrypoint = [ "${pkgs.trafficserver}/bin/traffic_server" ];
+			Entrypoint = [ "${custom_trafficserver}/bin/traffic_server" ];
 		};
 	};
 in image