+++ lib/bolt/core/src/dep/terraform/install_scripts/mod.rs
@@ -10,6 +10,7 @@ pub mod components;
 pub async fn gen(
 	ctx: &ProjectContext,
 	server: &Server,
+	all_servers: &HashMap<String, Server>,
 	k8s_infra: &terraform::output::K8sInfra,
 	tls: &terraform::output::Tls,
 ) -> Result<String> {
@@ -62,6 +63,8 @@ pub async fn gen(
 		script.push(components::docker()); // why do we need to install docker and cni plugins?
 		script.push(components::cni_plugins());
 		script.push(components::nomad(server));
+		script.push(components::envoy());
+		script.push(components::outbound_proxy(server, all_servers)?);
 
 		prometheus_targets.insert(
 			"nomad".into(),
@@ -75,7 +78,7 @@ pub async fn gen(
 	// MARK: ATS
 	if server.pool_id == "ats" {
 		script.push(components::docker());
-		script.push(components::traffic_server(ctx).await?);
+		script.push(components::traffic_server(ctx, server).await?);
 	}
 
 	// MARK: Common (post)