+++ svc/pkg/mm/worker/src/workers/lobby_create/mod.rs
@@ -738,12 +738,11 @@ async fn resolve_image_artifact_url(
 		backend::upload::Provider::Aws => s3_util::Provider::Aws,
 	};
 
-	match internal_unwrap_owned!(
+	let mm_lobby_delivery_method = internal_unwrap_owned!(
 		std::env::var("RIVET_MM_LOBBY_DELIVERY_METHOD").ok(),
 		"missing RIVET_MM_LOBBY_DELIVERY_METHOD"
-	)
-	.as_str()
-	{
+	);
+	match mm_lobby_delivery_method.as_str() {
 		"s3_direct" => {
 			tracing::info!("using s3 direct delivery");
 
@@ -776,20 +775,9 @@ async fn resolve_image_artifact_url(
 		"traffic_server" => {
 			tracing::info!("using traffic server delivery");
 
-			// HACK: Hardcode ATS IP since this will be replaced shortly
-			let ats_url = if region.name_id == "lnd-atl" {
-				"10.0.25.2"
-			} else if region.name_id == "lnd-fra" {
-				"10.0.50.2"
-			} else {
-				tracing::info!(?region.name_id);
-				internal_panic!("invalid region");
-			};
-
 			let upload_id = internal_unwrap!(upload.upload_id).as_uuid();
 			let addr = format!(
-				"http://{ats_url}:8080/s3-cache/{provider}/{namespace}-bucket-build/{upload_id}/image.tar",
-				ats_url = ats_url,
+				"http://127.0.0.1:8080/s3-cache/{provider}/{namespace}-bucket-build/{upload_id}/image.tar",
 				provider = heck::KebabCase::to_kebab_case(provider.as_str()),
 				namespace = util::env::namespace(),
 				upload_id = upload_id,